requirements.txt
ollama
edge-tts
pyttsx3
SpeechRecognition
customtkinter
pyautogui
pycaw
screen-brightness-control
requests
beautifulsoup4
duckduckgo-search
pyaudio
pygame
psutil
pygetwindow
pyperclip
pynput
wikipedia
PyPDF2
Pillow
schedule
